requirements.txt
requests
pandas
numpy
scipy
scikit-learn
python-slugify
leven
